chore: move files to dedicated paths
