+++ src/app/components/album-art/album-art.html
@@ -1,0 +1,5 @@
+<div class="album-art-container">
+  <div class="album-art-bg" [style.background-image]="src()"></div>
+  <div class="album-art-border"></div>
+  <div class="album-art-fg" [style.background-image]="src()"></div>
+</div>
