+++ src/app/components/album-art/album-art.scss
@@ -1,0 +1,56 @@
+
+.album-art-container {
+  position: relative;
+  width: fit-content;
+  --border-size: 10px;
+  --art-size: 212px;
+  --border-radius: 16px;
+  height: var(--art-size);
+  width: var(--art-size);
+  border-radius: var(--border-radius);
+  overflow: hidden;
+  box-shadow:
+          0px -0.5px 1px 0px rgba(255, 255, 255, 0.3) inset,
+          0px -0.5px 1px 0px rgba(255, 255, 255, 0.25) inset,
+          0px 1.5px 4px 0px rgba(0, 0, 0, 0.08) inset,
+          0px 1.5px 4px 0px rgba(0, 0, 0, 0.1) inset;
+}
+
+.album-art-bg {
+  background-size: var(--art-size);
+  background-repeat: no-repeat;
+  background-position: center;
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+}
+
+.album-art-border {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-image: linear-gradient(
+                  180deg,
+                  rgba(255, 255, 255, 0.2) 0%,
+                  rgba(255, 255, 255, 0) 41%,
+                  rgba(255, 255, 255, 0) 57%,
+                  rgba(255, 255, 255, 0.1) 100%
+  );
+  /*background-color: lightblue;*/
+}
+
+.album-art-fg {
+  background-size: var(--art-size);
+  background-repeat: no-repeat;
+  background-position: center;
+  position: absolute;
+  top: var(--border-size);
+  left: var(--border-size);
+  right: var(--border-size);
+  bottom: var(--border-size);
+  border-radius: calc(var(--border-radius) - var(--border-size));
+}
