+++ src/app/components/controls/play-pause/play-pause-btn.ts
@@ -1,0 +1,14 @@
+import { Component, inject, input, output } from "@angular/core";
+import { Metadata } from "../../../injectables/metadata";
+
+@Component({
+  selector: "play-pause-btn",
+  templateUrl: "./play-pause-btn.html",
+  styleUrl: "./play-pause-btn.scss",
+})
+export class PlayPauseBtn {
+  metadata = inject(Metadata);
+  resumeit = output();
+  pauseit = output();
+  isPaused = input.required<boolean>();
+}
