+++ src/app/components/controls/play-pause/play-pause.scss
@@ -1,0 +1,15 @@
+.buttonsRow {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 12px;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  border-radius: 14px;
+}
+
+.invisibleButton {
+  background: none;
+  border: none;
+  padding: 12px;
+}
