+++ src/app/components/controls/volume-slider/volume-slider.scss
@@ -1,0 +1,59 @@
+
+:host {
+  display: block;
+  width: 100%;
+}
+
+.volumeRow {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 12px;
+  padding: 12px 1rem;
+  border-radius: 14px;
+}
+
+.iconButton {
+  background: none;
+  border: none;
+  padding: 12px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.backArrow {
+  font-size: 28px;
+  color: rgba(255, 255, 255, 0.8);
+  line-height: 1;
+}
+
+.volumeSlider {
+  flex: 1;
+  height: 8px;
+  border-radius: 1.5rem;
+  appearance: none;
+  background: rgba(0, 0, 0, 0.12);
+  box-shadow:
+          0px -0.5px 1px 0px rgba(255, 255, 255, 0.3) inset,
+          0px -0.5px 1px 0px rgba(255, 255, 255, 0.25) inset,
+          0px 1.5px 4px 0px rgba(0, 0, 0, 0.08) inset,
+          0px 1.5px 4px 0px rgba(0, 0, 0, 0.1) inset;
+}
+
+.volumeSlider::-webkit-slider-thumb {
+  appearance: none;
+  width: 16px;
+  height: 16px;
+  border-radius: 50%;
+  background: rgba(255, 255, 255, 0.9);
+  cursor: pointer;
+  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.2);
+  margin-top: -4px;
+}
+
+.volumeSlider::-webkit-slider-runnable-track {
+  height: 8px;
+  border-radius: 1.5rem;
+}
