+++ src/app/components/currently-playing-material/currently-playing-material.html
@@ -1,0 +1,6 @@
+<div
+  class="coverImgBg"
+  style="background-image: {{ stateMetadata.urlCoverImage() }}"
+></div>
+<div class="scrim"></div>
+<ng-content></ng-content>
