+++ src/app/components/currently-playing-material/currently-playing-material.scss
@@ -1,0 +1,25 @@
+.coverImgBg {
+  position: fixed;
+  z-index: -2;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-size: cover;
+  filter: blur(90px);
+}
+
+.scrim {
+  position: fixed;
+  z-index: -1;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-size: 50px 50px;
+  background-repeat: repeat;
+  background-position: 0% 0%;
+  background-color: #222222;
+  mix-blend-mode: overlay;
+  backdrop-filter: blur(27px);
+}
