+++ src/app/components/currently-playing-material/currently-playing-material.ts
@@ -1,0 +1,11 @@
+import { Component, inject } from "@angular/core";
+import { Metadata } from "../../injectables/metadata";
+
+@Component({
+  selector: "currently-playing-material",
+  templateUrl: "./currently-playing-material.html",
+  styleUrl: "./currently-playing-material.scss",
+})
+export class CurrentlyPlayingMaterial {
+  stateMetadata = inject(Metadata);
+}
