+++ src/app/components/light-glass-outlined/light-glass-outlined.html
@@ -1,0 +1,3 @@
+<div class="innerContentsContainer">
+    <ng-content></ng-content>
+</div>
