+++ src/app/components/light-glass-outlined/light-glass-outlined.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "light-glass-outlined",
+  templateUrl: "./light-glass-outlined.html",
+  styleUrl: "./light-glass-outlined.scss",
+})
+export class LightGlassOutlined {}
