--- src/app/components/under-titlebar.component.ts
@@ -1,69 +1,0 @@
-import { Component, effect, input } from "@angular/core";
-import { getCurrentWindow } from "@tauri-apps/api/window";
-import { injectQuery } from "@tanstack/angular-query-experimental";
-import { listen, TauriEvent } from "@tauri-apps/api/event";
-import { getPlatform } from "../utils/platform";
-
-@Component({
-  selector: "app-under-titlebar-macos",
-  template: `
-    <div class="under-titlebar-macos">
-      <img
-        class="logo"
-        [src]="
-          theme() === 'light'
-            ? 'assets/header-logo-black.svg'
-            : 'assets/header-logo-white.svg'
-        "
-        alt=""
-        height="32"
-        width="32"
-      />
-    </div>
-  `,
-  styles: [
-    `
-      .under-titlebar-macos {
-        display: flex;
-        flex-direction: row;
-        justify-content: flex-end;
-      }
-
-      .logo {
-        opacity: 0.5;
-        mix-blend-mode: plus-lighter;
-      }
-    `,
-  ],
-})
-export class UnderTitlebarMacos {
-  theme = input<"light" | "dark">();
-}
-
-@Component({
-  selector: "app-under-titlebar",
-  imports: [UnderTitlebarMacos],
-  template: `
-    @if (currentPlatform === "macos") {
-      <app-under-titlebar-macos [theme]="theme.data() ?? 'dark'" />
-    } @else {
-      <div style="height: 32px"></div>
-    }
-  `,
-})
-export class UnderTitlebar {
-  currentPlatform = getPlatform();
-
-  theme = injectQuery(() => ({
-    queryKey: ["theme"],
-    queryFn: async () => (await getCurrentWindow().theme()) ?? "dark",
-  }));
-
-  _themeChange = effect((onCleanup) => {
-    let cleanup = () => {};
-    listen(TauriEvent.WINDOW_THEME_CHANGED, () => {
-      this.theme.refetch();
-    }).then((c) => (cleanup = c));
-    onCleanup(() => cleanup());
-  });
-}
