+++ src/app/components/under-titlebar/under-titlebar-macos.html
@@ -1,0 +1,13 @@
+<div class="under-titlebar-macos">
+  <img
+    class="logo"
+    [src]="
+      theme() === 'light'
+        ? 'assets/header-logo-black.svg'
+        : 'assets/header-logo-white.svg'
+    "
+    alt=""
+    height="32"
+    width="32"
+  />
+</div>
