--- src/app/views/currently-playing.component.ts
@@ -1,402 +1,0 @@
-import { Component, computed, inject, input, signal } from "@angular/core";
-import { injectSelector } from "@reduxjs/angular-redux";
-import { RootState } from "../store";
-import { pause, play, resume, seek } from "../services/music";
-import { pickSong } from "../services/fs";
-import { Metadata } from "../injectables/metadata";
-import { injectMutation } from "@tanstack/angular-query-experimental";
-import { PlayPause } from "../components/controls/play-pause";
-import { VolumeSlider } from "../components/controls/volume-slider";
-import { zeroPad } from "../utils/strings";
-import { ControlType } from "../types/currently-playing";
-
-@Component({
-  selector: "currently-playing-material",
-  template: `
-    <div
-      class="coverImgBg"
-      style="background-image: {{ stateMetadata.urlCoverImage() }}"
-    ></div>
-    <div class="scrim"></div>
-    <ng-content></ng-content>
-  `,
-  styles: [
-    `
-      .coverImgBg {
-        position: fixed;
-        z-index: -2;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        background-size: cover;
-        filter: blur(90px);
-      }
-
-      .scrim {
-        position: fixed;
-        z-index: -1;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        background-size: 50px 50px;
-        background-repeat: repeat;
-        background-position: 0% 0%;
-        background-color: #222222;
-        mix-blend-mode: overlay;
-        backdrop-filter: blur(27px);
-      }
-    `,
-  ],
-})
-class CurrentlyPlayingMaterial {
-  stateMetadata = inject(Metadata);
-}
-
-@Component({
-  selector: "album-art",
-  template: `
-    <div class="album-art-container">
-      <div class="album-art-bg" [style.background-image]="src()"></div>
-      <div class="album-art-border"></div>
-      <div class="album-art-fg" [style.background-image]="src()"></div>
-    </div>
-  `,
-  styles: [
-    `
-      .album-art-container {
-        position: relative;
-        width: fit-content;
-        --border-size: 10px;
-        --art-size: 212px;
-        --border-radius: 16px;
-        height: var(--art-size);
-        width: var(--art-size);
-        border-radius: var(--border-radius);
-        overflow: hidden;
-        box-shadow:
-          0px -0.5px 1px 0px rgba(255, 255, 255, 0.3) inset,
-          0px -0.5px 1px 0px rgba(255, 255, 255, 0.25) inset,
-          0px 1.5px 4px 0px rgba(0, 0, 0, 0.08) inset,
-          0px 1.5px 4px 0px rgba(0, 0, 0, 0.1) inset;
-      }
-
-      .album-art-bg {
-        background-size: var(--art-size);
-        background-repeat: no-repeat;
-        background-position: center;
-        position: absolute;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-      }
-
-      .album-art-border {
-        position: absolute;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-        background-image: linear-gradient(
-          180deg,
-          rgba(255, 255, 255, 0.2) 0%,
-          rgba(255, 255, 255, 0) 41%,
-          rgba(255, 255, 255, 0) 57%,
-          rgba(255, 255, 255, 0.1) 100%
-        );
-        /*background-color: lightblue;*/
-      }
-
-      .album-art-fg {
-        background-size: var(--art-size);
-        background-repeat: no-repeat;
-        background-position: center;
-        position: absolute;
-        top: var(--border-size);
-        left: var(--border-size);
-        right: var(--border-size);
-        bottom: var(--border-size);
-        border-radius: calc(var(--border-radius) - var(--border-size));
-      }
-    `,
-  ],
-})
-export class AlbumArt {
-  src = input<string | null>();
-}
-
-@Component({
-  selector: "app-now-playing",
-  imports: [CurrentlyPlayingMaterial, AlbumArt, PlayPause, VolumeSlider],
-  template: `
-    <currently-playing-material>
-      <div class="container">
-        @if (!openFileMutation.data()) {
-          <button (click)="openFileMutation.mutate()">Play</button>
-        }
-
-        @if (
-          openFileMutation.data() &&
-          stateMetadata.metadata.isSuccess() &&
-          openFileMutation.isSuccess()
-        ) {
-          <album-art [src]="stateMetadata.urlCoverImage()"></album-art>
-          <div class="infoContainer">
-            <div class="textContainer">
-              <h1 class="title">
-                {{
-                  stateMetadata.metadata.data()?.tags?.TrackTitle ??
-                    "Unknown Title"
-                }}
-              </h1>
-              <p class="artist">
-                {{
-                  stateMetadata.metadata.data()?.tags?.Artist ??
-                    "Unknown Artist"
-                }}
-              </p>
-              <p class="album">
-                {{
-                  stateMetadata.metadata.data()?.tags?.Album ?? "Unknown Album"
-                }}
-              </p>
-            </div>
-
-            <div class="progressContainer">
-              <div class="progressTextContainer">
-                <p class="progressText">{{ currentTimeFormatted() }}</p>
-                <p class="progressText">{{ totalTimeFormatted() }}</p>
-              </div>
-              <!-- TODO: Add left and right keybinds -->
-              <input
-                type="range"
-                class="progressBar"
-                [value]="currentSecs() ?? 0"
-                [max]="totalSecs() ?? 0"
-                min="0"
-                step="1"
-                [style.--progress]="progressPercent() + '%'"
-                (input)="seekFromInput($event)"
-              />
-            </div>
-
-            <div>
-              @if (controls() === 'volume-change') {
-                <app-volume-slider
-                  [volume]="playingMetadata().volume"
-                  (backClick)="controls.set('play-pause')"
-                />
-              } @else {
-                <app-play-pause
-                  (pauseit)="pause()"
-                  (resumeit)="resume()"
-                  (volumeClick)="controls.set('volume-change')"
-                  [isPaused]="playingMetadata().paused"
-                />
-              }
-            </div>
-          </div>
-        }
-      </div>
-    </currently-playing-material>
-  `,
-  styles: [
-    `
-      .container {
-        padding: 24px;
-        display: flex;
-        flex-direction: row;
-        overflow: hidden;
-        justify-content: center;
-        align-items: center;
-        /* 32 because of top area */
-        height: calc(100vh - 32px);
-        box-sizing: border-box;
-      }
-
-      .infoContainer {
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-        justify-content: center;
-        align-items: center;
-        padding: 0 32px;
-        text-align: center;
-        width: 1px;
-        flex-grow: 1;
-      }
-
-      .textContainer {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-        justify-content: center;
-        align-items: center;
-        width: 100%;
-      }
-
-      .title {
-        font-size: 24px;
-        font-weight: bold;
-        margin: 0;
-        color: white;
-        width: 100%;
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-      }
-
-      .artist {
-        font-size: 16px;
-        margin: 0;
-        color: white;
-        opacity: 0.8;
-      }
-
-      .album {
-        font-size: 16px;
-        margin: 0;
-        color: white;
-        opacity: 0.8;
-      }
-
-      .progressContainer {
-        width: 100%;
-        padding: 8px 0;
-        flex-direction: column;
-        gap: 8px;
-      }
-
-      .progressTextContainer {
-        display: flex;
-        justify-content: space-between;
-      }
-
-      .progressText {
-        font-size: 14px;
-        color: white;
-        opacity: 0.5;
-        margin: 0;
-      }
-
-      .progressBar {
-        border-radius: 1.5rem;
-        width: 100%;
-        box-shadow:
-          0px -0.5px 1px 0px rgba(255, 255, 255, 0.3) inset,
-          0px -0.5px 1px 0px rgba(255, 255, 255, 0.25) inset,
-          0px 1.5px 4px 0px rgba(0, 0, 0, 0.08) inset,
-          0px 1.5px 4px 0px rgba(0, 0, 0, 0.1) inset;
-        background-blend-mode: plus-darker;
-        height: 0.5rem;
-        appearance: none;
-        background: rgba(0, 0, 0, 0.12);
-        cursor: pointer;
-      }
-
-      .progressBar::-webkit-slider-thumb {
-        appearance: none;
-        width: 0;
-        height: 0;
-      }
-
-      .progressBar::-webkit-slider-runnable-track {
-        height: 0.5rem;
-        border-radius: 1.5rem;
-        background: linear-gradient(
-          to right,
-          rgba(255, 255, 255, 0.8) 0%,
-          rgba(255, 255, 255, 0.8) var(--progress, 0%),
-          transparent var(--progress, 0%),
-          transparent 100%
-        );
-      }
-    `,
-  ],
-})
-export class CurrentlyPlaying {
-  playingMetadata = injectSelector((state: RootState) => state.tauri);
-  stateMetadata = inject(Metadata);
-
-  controls = signal<ControlType>("play-pause");
-
-  openFileMutation = injectMutation(() => ({
-    mutationKey: ["openFile"],
-    mutationFn: async () => {
-      const path = await pickSong();
-      if (!path) {
-        return;
-      }
-      await play(path);
-      return true;
-    },
-  }));
-
-  totalSecs = computed(() => {
-    return this.playingMetadata().duration?.secs;
-  });
-
-  currentSecs = computed(() => {
-    return this.playingMetadata().position?.secs;
-  });
-
-  progressPercent = computed(() => {
-    const current = this.currentSecs() ?? 0;
-    const total = this.totalSecs() ?? 0;
-    if (total === 0) return 0;
-    return (current / total) * 100;
-  });
-
-  totalTime = computed(() => {
-    // Format as `hh:mm:ss` but only show minutes and hours if they are non-zero
-    const duration = this.playingMetadata().duration?.secs;
-    if (!duration) {
-      return {
-        hours: 0,
-        minutes: 0,
-        seconds: 0,
-      };
-    }
-    // Format
-    const hours = Math.floor(duration / 3600);
-    const minutes = Math.floor(duration / 60) % 60;
-    const seconds = duration % 60;
-    return {
-      hours,
-      minutes,
-      seconds,
-    };
-  });
-
-  totalTimeFormatted = computed(() => {
-    const duration = this.totalTime();
-    return `${duration.hours ? duration.hours + ":" : ""}${zeroPad(duration.minutes, 2)}:${zeroPad(duration.seconds, 2)}`;
-  });
-
-  currentTimeFormatted = computed(() => {
-    // Format as `hh:mm:ss` but only if `totalTime` has an hour or minute value
-    const totalDuration = this.totalTime();
-    const duration = this.playingMetadata().position?.secs;
-    if (!duration) {
-      return "0:00";
-    }
-    // Format
-    const hours = Math.floor(duration / 3600);
-    const minutes = Math.floor(duration / 60) % 60;
-    const seconds = duration % 60;
-    const showHours = totalDuration.hours > 0 || hours > 0;
-    const showMinutes = totalDuration.minutes > 0 || minutes > 0;
-    return `${showHours ? hours + ":" : ""}${showMinutes ? zeroPad(minutes, 2) : ""}:${zeroPad(seconds, 2)}`;
-  });
-
-  resume = resume;
-  pause = pause;
-
-  seekFromInput(event: Event) {
-    const input = event.target as HTMLInputElement;
-    const value = parseFloat(input.value);
-    seek(value);
-  }
-}
