+++ src/app/views/settings/settings.html
@@ -1,0 +1,8 @@
+<div class="container">
+  <light-glass-outlined>
+    <div class="innerContents">
+      <h1 class="title">Settings</h1>
+      <hr class="hr" />
+    </div>
+  </light-glass-outlined>
+</div>
